+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Fix: Remove onActivityPreCreated call in favor of onActivityCreated (#1661)
 * Fix: Do not crash if SENSOR_SERVICE throws (#1655)
 * Feat: Add support for async methods in Spring MVC (#1652)
 * Feat: Add secondary constructor taking IHub to SentryOkHttpInterceptor (#1657)